Criação de consultas (ajustes em classes para relacionamento one to many e many to one

+++ src/main/java/com/example/sistemaDeReservasDeConsultas/service/ConsultaServiceImpl.java
@@ -1,6 +1,6 @@
 package com.example.sistemaDeReservasDeConsultas.service;
 
-import com.example.sistemaDeReservasDeConsultas.entity.ConsultaModel;
+import com.example.sistemaDeReservasDeConsultas.model.Consulta;
 import com.example.sistemaDeReservasDeConsultas.repository.IConsultaRepository;
 import org.springframework.stereotype.Service;
 
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-public class ConsultaServiceImpl implements IConsultaService<ConsultaModel> {
+public class ConsultaServiceImpl implements IService<Consulta> {
     private final IConsultaRepository repository;
 
     public ConsultaServiceImpl(IConsultaRepository repository) {
@@ -16,40 +16,37 @@ public class ConsultaServiceImpl implements IConsultaService<ConsultaModel> {
     }
 
     @Override
-    public List<ConsultaModel> getAll() {
+    public Consulta add(Consulta consulta) {
+        if (consulta != null) { return repository.save(consulta); }
+        return new Consulta();
+    }
+
+    @Override
+    public List<Consulta> getAll() {
         return repository.findAll();
     }
 
     @Override
-    public Optional<ConsultaModel> getById(Long id) {
+    public Optional<Consulta> getById(Long id) {
         return repository.findById(id);
     }
 
     @Override
-    public ConsultaModel add(ConsultaModel consulta) {
-        if (consulta != null) {
-            return repository.save(consulta);
-        }
-        return new ConsultaModel();
+    public void remove(Long id) {
+        if (getById(id).isPresent()) { repository.deleteById(id); }
+        else {
+            throw new IllegalStateException("Repositório nulo"); }
     }
 
     @Override
-    public void update(ConsultaModel consulta) {
-        if (consulta == null) {
-            throw new IllegalStateException("Consulta nulo");
-        } else if (consulta != null && repository.findById(consulta.getId()).isPresent()) {
+    public void update(Consulta consulta) {
+        if (consulta == null) { throw new IllegalStateException("Consulta nulo"); }
+        else if (consulta != null && repository.findById(consulta.getId()).isPresent()) {
             repository.saveAndFlush(consulta);
         }
     }
 
-    @Override
-    public void remove(Long id) {
-        if (getById(id).isPresent()) {
-            repository.deleteById(id);
-        } else {
-            throw new IllegalStateException("Repositório nulo");
-        }
-    }
+
 }
